fix: programmatic usage throws error on import
